requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
betterproto==1.2.5
boid-model @ git+ssh://git@github.com/BOID-Smart-Assistent/BOID-model.git@a6f2d95569bae96a8aba0cc1c7d7b7833a0b9b26
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
Deprecated==1.2.15
dirtyjson==1.0.8
distro==1.9.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
grpclib==0.4.7
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
jiter==0.8.0
joblib==1.4.2
llama-cloud==0.1.6
llama-index==0.12.2
llama-index-agent-openai==0.4.0
llama-index-cli==0.4.0
llama-index-core==0.12.2
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.3.2
llama-index-multi-modal-llms-openai==0.3.0
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.0
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.16
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
ollama==0.4.2
openai==1.56.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
pydantic==2.9.2
pydantic_core==2.23.4
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stringcase==1.2.0
striprtf==0.0.26
tenacity==8.5.0
tiktoken==0.8.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websockets==14.1
wrapt==1.17.0
yarl==1.18.3
